Move iOS image selection logic to separate class
